fix: 가격 관련 컬럼의 타입 BigDecimal로 변경

+++ src/main/java/com/example/monghyang/domain/brewery/joy/dto/ReqJoyDto.java
@@ -6,6 +6,8 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.springframework.web.multipart.MultipartFile;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 public class ReqJoyDto {
@@ -16,6 +18,6 @@ public class ReqJoyDto {
     @NotNull(message = "체험의 설명 등 상세 정보를 입력해주세요.")
     private String detail;
     @NotNull(message = "체험의 1인 당 정가 정보를 정수로 입력해주세요.")
-    private Integer origin_price;
+    private BigDecimal origin_price;
     private MultipartFile image;
 }
